+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -12,6 +12,7 @@ public class ProjectManagerConst {
     // Application Properties
     public final static String PM_ADMIN_GROUP_PROPERTY = "project-manager-admin";
     public final static String JWT_GROUPS_CLAIM_PROPERTY = "jwt.groups.claim";
+    public final static String APP_SECURITY_ENABLED_PROPERTY = "app.security.enabled";
 
     // Environment Variables
     public final static String BK_USER_GROUP_PREFIX = "BK_USER_GROUP_PREFIX";
@@ -23,6 +24,7 @@ public class ProjectManagerConst {
     public final static String HEAD_SV = "${";
     public final static String BOTTOM_SV = "}";
     public final static String JWT_GROUPS_CLAIM_PROPERTY_SV = HEAD_SV + JWT_GROUPS_CLAIM_PROPERTY + BOTTOM_SV;
+    public final static String APP_SECURITY_ENABLED_PROPERTY_SV = HEAD_SV + APP_SECURITY_ENABLED_PROPERTY + ":true" + BOTTOM_SV;
     public final static String BK_USER_GROUP_PREFIX_SV = HEAD_SV + BK_USER_GROUP_PREFIX + BOTTOM_SV;
     public final static String BK_USER_GROUP_SUFFIX_SV = HEAD_SV + BK_USER_GROUP_SUFFIX + BOTTOM_SV;
     public final static String BK_ADMIN_GROUP_PREFIX_SV = HEAD_SV + BK_ADMIN_GROUP_PREFIX + BOTTOM_SV;
@@ -30,5 +32,7 @@ public class ProjectManagerConst {
 
     // Others
     public final static String BRIDGEHEAD_CLAIM = "bridgehead";
+    public final static String TEST_EMAIL = "test@project-manager.com";
+    public final static String TEST_BRIDGEHEAD = "bridgehead-test";
 
 }